backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.10.0
pydantic-ai==0.0.14
python-multipart==0.0.6
python-dotenv==1.0.0
mangum==0.17.0
